requirements.txt
tensorflow==2.21.0
tensorflowjs
numpy
google-cloud-storage
